yolo_kitti_final/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=0.11.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tqdm>=4.65.0
Pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
kaggle>=1.6.0
